% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limDis.R
\name{limDisAlt}
\alias{limDisAlt}
\title{Sample from the limit distribution under the alternative.}
\usage{
limDisAlt(B = 1000, r, s, distMat, p = 1)
}
\arguments{
\item{B}{Number of samples to generate.}

\item{r, s}{Number of counts giving the two samples.}

\item{distMat}{Distance matrix.}

\item{p}{Cost exponent. Defaults to 1.}
}
\value{
A vector of samples.
}
\description{
Sample from the limit distribution under the alternative.
}

