% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limDis.R
\name{limDisAltBoot}
\alias{limDisAltBoot}
\title{m-out-of-n Bootstrap for the limiting distribution.}
\usage{
limDisAltBoot(r, s, distMat, B = 1000, p = 1, gamma = 0.9)
}
\arguments{
\item{r, s}{Vectors of counts giving the two samples.}

\item{distMat}{Distance matrix.}

\item{B}{The number of samples to generate. Defaults to 1000.}

\item{p}{Cost exponent. Defaults to 1.}

\item{gamma}{m = n^gamma. Defaults to 0.9.}
}
\value{
A sample from the limiting distribution.
}
\description{
m-out-of-n Bootstrap for the limiting distribution.
}

