% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limDis.R
\name{limDisNull}
\alias{limDisNull}
\title{Sample from the limiting distribution under the null.}
\usage{
limDisNull(B = 500, r, distMat, p = 1)
}
\arguments{
\item{B}{number of samples to generate. Defaults to 500.}

\item{r}{vector of probabilities in the original problem.}

\item{distMat}{distance matrix in the original problem.}

\item{p}{cost exponent. Defaults to 1.}
}
\value{
A vector of samples.
}
\description{
Sample from the limiting distribution under the null.
}

