% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limDis.R
\name{limDisNullGrid}
\alias{limDisNullGrid}
\title{Sample from the limiting distribution under the null when the underlying
space is a grid.}
\usage{
limDisNullGrid(B = 500, r, p = 1)
}
\arguments{
\item{B}{Number of bootstrap samples to generate. Defaults to 500.}

\item{r}{vector of probabilities in the original problem. Is interpreted as
a square matrix.}

\item{p}{cost exponent.}
}
\value{
A vector of samples.
}
\description{
Sample from the limiting distribution under the null when the underlying
space is a grid.
}

