% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassDist.R
\name{wassDist}
\alias{wassDist}
\title{Compute the Wasserstein distance between to finite distributions.}
\usage{
wassDist(a, b, distMat, p = 1)
}
\arguments{
\item{a, b}{Vectors representing probability distributions.}

\item{distMat}{Cost matrix.}

\item{p}{cost exponent.}
}
\value{
The Wasserstein distance.
}
\description{
Compute the Wasserstein distance between to finite distributions.
}

